// rest-countries-theme-switcher/src/types/Country.ts
export interface Country {
  name: {
    common: string
    official: string
    nativeName?: {
      [key: string]: {
        official: string
        common: string
      }
    }
  }

  tld?: string[]

  cca2: string
  cca3: string
  ccn3?: string

  capital?: string[]

  region: string
  subregion?: string

  languages?: {
    [key: string]: string
  }

  currencies?: {
    [key: string]: {
      name: string
      symbol?: string
    }
  }

  population: number

  flags: {
    png: string
    svg: string
    alt?: string
  }

  borders?: string[]

  maps?: {
    googleMaps: string
    openStreetMaps: string
  }

  continents?: string[]

  timezones?: string[]
}